+++ index.html
@@ -234,8 +234,8 @@
       </section>
     </template>
 
-  <!-- <script src="/js/functions.js" ></script> -->
-  <script src="/js/main.js" type="module"></script>
+  <script src="./js/functions_2-33/functions.js" ></script>
+  <script src="./js/main.js" type="module"></script>
 
 </body>
