+++ js/create-photos/create-data.js
@@ -1,9 +1,7 @@
-// точка входа - модуль, который связывает другие модули
-
-import { getDataArrays } from './data.js';
 import { getRandomInteger, getRandomElements } from './utils.js';
-import { generatePhotoId, generatePhotoUrl } from './get-id.js';
-import { getAllComments } from './comments-photo.js';
+import { getDataArrays } from './data.js';
+import { generatePhotoId } from './create-photos/create-id.js';
+import { getAllComments } from './create-photos/comments-photo.js';
 
 /**
  * Деструктуризация импортированных данных
@@ -11,15 +9,17 @@ import { getAllComments } from './comments-photo.js';
 const { MAX_PHOTOS, LIKES, DESCRIPTIONS, NAMES, MESSAGES } = getDataArrays();
 
 /** Функция для создания объекта с описанием фотографии
-  * @param {int} id - идентификатор фотографии
+  * @param {number} id - идентификатор фотографии
   * @param {string} url - ссылка на фотографию
   * @param {string} description - описание фотографии
-  * @param {int} likes - количество лайков
-  * @param {Array} generateCommentsToPhoto() - массив комментариев
+  * @param {number} likes - количество лайков
+  * @param {int} getAllComments() - массив комментариев
+  * @param {Array} getPhoto
  * Структура каждого объекта должна быть следующей:
    - id, число — идентификатор опубликованной фотографии. Это число от 1 до 25. Идентификаторы не должны повторяться.
-   - url, строка — адрес картинки вида photos/{{i}}.jpg, где {{i}} — это число от 1 до 25. Адреса картинок не должны повторяться.
+   - url, строка — адрес картинки вида photos/{{i}}.jpg, где {{i}} — это число от 1 до 25, i = getId. Адреса картинок не должны повторяться.
    - description, строка — описание фотографии. Описание придумайте самостоятельно.
+   - likes
    - comments, массив объектов — список комментариев, оставленных другими пользователями к этой фотографии. Количество комментариев к каждой фотографии — случайное число от 0 до 30. Все комментарии генерируются случайным образом. Пример описания объекта с комментарием:
 */
 const getPhotoUsers = () => {
