+++ js/data.js
@@ -1,4 +1,3 @@
-
 // модуль, который создаёт данные - константы и массивы
 
 export const getDataArrays = () => {
