+++ js/main.js
@@ -1,7 +1,7 @@
 // точка входа - модуль, который связывает другие модули
 
-import { getAllPhotoUsers } from './create-data.js';
-import { createPhotos } from './thumbnails.js';
+import { getAllPhotoUsers } from './create-photos/create-data.js';
+import { createPhotos } from './create-photos/thumbnails.js';
 
 const photosArray = getAllPhotoUsers();
 createPhotos(photosArray);
